+++ src/components/molecules/NetworkBanner.tsx
@@ -1,21 +1,15 @@
 import React, { ReactElement, useEffect, useState } from 'react'
-import styles from './AnnouncementBanner.module.css'
-import Markdown from '../atoms/Markdown'
 import { useWeb3 } from '../../providers/Web3'
 import { addCustomNetwork, NetworkObject } from '../../utils/web3'
 import { getOceanConfig } from '../../utils/ocean'
 import { getProviderInfo } from 'web3modal'
 import { useOcean } from '../../providers/Ocean'
 import { useSiteMetadata } from '../../hooks/useSiteMetadata'
-import Button from '../atoms/Button'
+import AnnouncementBanner, {
+  AnnouncementAction
+} from '../atoms/AnnouncementBanner'
 
-export interface AnnouncementAction {
-  name: string
-  style?: string
-  handleAction: () => void
-}
-
-const network: NetworkObject = {
+const networkMatic: NetworkObject = {
   chainId: 137,
   name: 'Matic Network',
   urlList: [
@@ -24,22 +18,17 @@ const network: NetworkObject = {
   ]
 }
 
-export default function AnnouncementBanner({
-  graphSynched
-}: {
-  graphSynched: boolean
-}): ReactElement {
+export default function NetworkBanner(): ReactElement {
   const { web3Provider } = useWeb3()
   const { config, connect } = useOcean()
-  const { announcement, warning } = useSiteMetadata()
+  const { announcement } = useSiteMetadata()
 
   const [text, setText] = useState<string>(announcement.main)
-  const [secondaryText, setSecondaryText] = useState<string>()
   const [action, setAction] = useState<AnnouncementAction>()
 
   const addCustomNetworkAction = {
     name: 'Add custom network',
-    handleAction: () => addCustomNetwork(web3Provider, network)
+    handleAction: () => addCustomNetwork(web3Provider, networkMatic)
   }
   const switchToPolygonAction = {
     name: 'Switch to Polygon',
@@ -60,12 +49,6 @@ export default function AnnouncementBanner({
     setText(announcement.polygon)
     setAction(undefined)
   }
-
-  useEffect(() => {
-    graphSynched
-      ? setSecondaryText(undefined)
-      : setSecondaryText(warning.graphNotSynched)
-  }, [graphSynched])
 
   useEffect(() => {
     if (!web3Provider && !config) return
@@ -99,21 +82,5 @@ export default function AnnouncementBanner({
     }
   }, [web3Provider, config, announcement])
 
-  return (
-    <div className={styles.container}>
-      <div className={styles.banner}>
-        {text && (
-          <Markdown
-            className={styles.text}
-            text={secondaryText ? secondaryText + ' ' + text : text}
-          />
-        )}
-        {action && (
-          <Button style="text" size="small" onClick={action.handleAction}>
-            {action.name}
-          </Button>
-        )}
-      </div>
-    </div>
-  )
+  return <AnnouncementBanner text={text} action={action} />
 }
